+++ colors/meowsoot-moon.lua
@@ -1,0 +1,1 @@
+require("meowsoot").load({ style = "moon" })